+++ generators/csharp/sdk/CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.5.0 - 2024-07-30]
+
+- Feature: Add support for `uint`, `ulong`, and `float` types.
+- Internal: Bump to IRv53.
+
 ## [0.4.0 - 2024-07-30]
 
 - Feature: Add support for `allow-multiple` query parameters, which are sent in the `explode` format.